move processing to process_transcript
