+++ src/analyse_transcripts.py
@@ -4,16 +4,16 @@ import seaborn as sns
 import numpy as np
 import pandas as pd
 
-from datetime import datetime
 
+from datetime import datetime
 from tqdm import tqdm
 
-from utils import load_one_transcript, load_all_transcripts
-from predictDA import make_annotated_transcript
 
 import config
 
 from mappings import get_id2tag, get_tag2full_label
+
+from topics import add_topics_to_dfs
 
 def make_adjacency_matrix(tag_sequences, n, normalise = "next"):
 
@@ -47,42 +47,6 @@ def make_adjacency_matrix(tag_sequences, n, normalise = "next"):
 
     return adj_matrix
 
-def timestamp_to_datetime(timestamp):
-    return datetime.strptime(timestamp, "%H:%M:%S")
-
-
-def get_fractional_time(current_timestamp, biggest_time):
-    try:
-        current_seconds = (biggest_time - timestamp_to_datetime(current_timestamp)).total_seconds()
-        total_seconds = (biggest_time - timestamp_to_datetime("00:00:00")).total_seconds()
-        return 1 - current_seconds/total_seconds
-    except:
-        return current_timestamp
-
-def enhance_transcript_df(df):
-    #calc relative time
-    biggest_time = timestamp_to_datetime(df.loc[df.shape[0] - 1, "timestamp"])
-    df["relative_time"] = (
-        df["timestamp"].apply(get_fractional_time, args = (biggest_time, )))
-
-    #explicitly find speaker change positions
-    speaker_change = np.zeros(df.shape[0], dtype=np.int32)
-    speakers = df["speaker"]
-    for i, speaker in speakers[:-1].iteritems():
-        if speaker != speakers[i + 1]:
-            speaker_change[i + 1] = 1
-    df["speaker_change"] = speaker_change
-
-    # length of utterances in seconds
-    timestamps = df["timestamp"]
-    lengths = np.zeros(df.shape[0])
-    for i, timestamp in timestamps[:-1].iteritems():
-        next_timestamp = timestamps[i + 1]
-        lengths[i] = (timestamp_to_datetime(next_timestamp) - timestamp_to_datetime(timestamp)).total_seconds()
-    df["utterance_length"] = lengths
-
-    return df
-
 if __name__ == "__main__":
 
     max_nr_utterances = config.data["max_nr_utterances"]
@@ -101,7 +65,9 @@ if __name__ == "__main__":
     transcript_dfs = [make_annotated_transcript(t) for t in tqdm(transcripts)]
     for transcript_df in transcript_dfs:
         enhance_transcript_df(transcript_df)
-    merged_transcript_dfs = pd.concat(transcript_dfs)
+    #merged_transcript_dfs = pd.concat(transcript_dfs)
+
+    transcript_dfs = add_topics_to_dfs(transcript_dfs)
 
     #%%
     tag_sequences = [t_df["da_label"] for t_df in transcript_dfs]
